Extract list content in directories and update file paths
