+++ client/src/App.js
@@ -1,6 +1,6 @@
 import React, { useEffect } from "react"
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
-import Today from './components/Today'
+import Today from './components/day/Today'
 import Cover from "./components/cover/Cover"
 import EditProfile from "./components/EditProfile";
 import AppState from './context/application/AppState'
