+++ client/src/components/day/Today.js
@@ -1,12 +1,12 @@
 import React, {Fragment, useContext, useEffect} from 'react';
-import Weather from './Weather';
-import Day from './Day';
-import Datepicker from './Datepicker'
-import Content from './Content'
-import User from './User'
-import Alert from "../components/Alert"
+import Weather from '../Weather';
+import Day from '../Day';
+import Datepicker from '../Datepicker'
+import Content from './content/Content'
+import User from '../User'
+import Alert from "../Alert"
 
-import AuthContext from "../context/authentication/AuthContext";
+import AuthContext from "../../context/authentication/AuthContext";
 
 
 const Today = (props) => {
